bash scripts/bigquery/eviction/requirements.txt
pyyaml==3.12
apache-beam==2.16.0  
avro==1.9.1   
boto==2.49.0  
cachetools==3.1.0   
certifi==2019.3.9
chardet==3.0.4   
crcmod==1.7
Cython==0.29.28
dill==0.3.0   
docopt==0.6.2   
enum34==1.1.6   
fastavro==0.21.24 
fasteners==0.15
funcsigs==1.0.2
future==0.18.2
futures==3.2.0
google-api-core==1.14.0
google-apitools==0.5.28
google-auth==1.6.3
google-auth-oauthlib==0.4.0
google-cloud-bigquery==1.11.4
google-cloud-bigtable==1.0.0
google-cloud-core==1.0.2
google-cloud-datastore==1.7.4
google-cloud-pubsub==1.0.2
google-cloud-storage==1.16.2
google-compute-engine==2.8.13
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
googledatastore==7.0.2
grpc-google-iam-v1==0.12.3
grpcio==1.25.0
hdfs==2.5.8
httplib2==0.12.0
idna==2.8
linecache2==1.0.0
mock==2.0.0
monotonic==1.5
numpy==1.16.6
oauth2client==3.0.0
oauthlib==3.0.2
pandas==0.24.2
pandas-gbq==0.5.0
pbr==5.4.3
prometheus-client==0.6.0
proto-google-cloud-datastore-v1==0.90.4
protobuf==3.7.1
pyasn1==0.4.5 
pyasn1-modules==0.2.5
pydot==1.4.1
pymongo==3.9.0
pyparsing==2.4.5
python-dateutil==2.8.0
pytz==2019.1
PyVCF==0.6.8
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
setuptools==41.0.1
simplejson==3.16.0
six==1.12.0
traceback2==1.4.0
typing==3.6.6
unittest2==1.1.0
urllib3==1.24.2
wheel==0.33.1
yml==0.0.1
